tests and example test
