+++ cassowary-rs-variable/Cargo.toml
@@ -7,4 +7,4 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cassowary-rs = { path = "../cassowary-rs" }
+cassowary = { path = "../cassowary-rs" }
